% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxtablefunction.R
\name{cox_table}
\alias{cox_table}
\title{cox propotional hazard ratio}
\usage{
cox_table(yr, outcome, group, dat, digit = 3)
}
\arguments{
\item{yr}{Survival duration}

\item{outcome}{Survival outcome}

\item{group}{Interested variable}
}
\description{
This function allows you to calculate HR, CI, p-value by Cox
}
\examples{
cox_table(survival duration, survival outcome, variable intrested, data)
}
\keyword{coxtable}
